feat: switch to `u32` for easier types

+++ agent_api_http/src/v0/issuance/public_offers.rs
@@ -17,7 +17,7 @@ use std::sync::Arc;
 pub struct PublicOfferStatusDto {
     pub id: String,
     pub template_id: String,
-    pub amount_issued: u64,
+    pub amount_issued: u32,
     pub status: PublicOfferStatus,
 }
